Feature/caf subscription (#1012)
* feature(children/new) : update submit button text based on form path and ENV['CAF_SUBSCRIPTION']

* feature(children) : enhance registration flow with CAF subscription handling

* feature(children) caf subscription optimization

* feature(children) : avoid sending messages depending on caf subscription

* feature(workshops) :update wording and add address supplement field to workshops and update related forms

* feature(children) : add safe navigation to prevent nil errors in late support warning

* feature(children) : conditionally require email input for CAF registration

* fix(children_controller) : correct typo in parent1 last name reference

* fix(child_create_service) : update SMS sending logic to include CAF subscription check

* Feature/subscription reminder (#1013)

* feature(caf_subscription_reminder) : add CAF subscription reminder job and service, and integrate with Typeform webhook

* fix(program_caf_subscription_reminder) : update service reference and correct message parameter keys in CAF subscription reminder job and service

* fix(typeform_controller, new.html.erb) : improve error handling and update CAF subscription check logic

+++ app/controllers/children_controller.rb
@@ -37,6 +37,15 @@ class ChildrenController < ApplicationController
       build_child_for_form
       @child.build_children_source(source_id: children_source_params&.dig(:source_id), details: children_source_params&.dig(:details), registration_department: children_source_params&.dig(:registration_department))
       render action: :new
+    elsif current_registration_origin == 2 && ENV['CAF_SUBSCRIPTION'].present?
+      redirect_to created_child_path(
+        child_support_id: @child.child_support.id,
+        current_child_name: @child.first_name,
+        parent1_last_name: @child.parent1.last_name,
+        email: @child.parent1.email,
+        current_child_months: @child.months,
+        sc: @child.parent1.security_code
+      )
     elsif service.parent1_target_profile || current_registration_origin != 4
       redirect_to created_child_path(sms_url_form: service.sms_url_form, children_under_four_months: service.children_under_four_months, youngest_child_under_twenty_four_months: service.youngest_child_under_twenty_four_months)
     else
@@ -82,11 +91,15 @@ class ChildrenController < ApplicationController
       @widget = false
       @new_link = new_pmi_registration_path
     when 2
-      session.delete(:registration_origin)
-      @message = I18n.t(support_wait_time_message, typeform_url: params[:sms_url_form])
-      @again = false
-      @widget = false
-      @new_link = new_caf_registration_path
+      if ENV['CAF_SUBSCRIPTION'].present?
+        @caf_subscription_form = true
+      else
+        session.delete(:registration_origin)
+        @message = I18n.t(support_wait_time_message, typeform_url: params[:sms_url_form])
+        @again = false
+        @widget = false
+        @new_link = new_caf_registration_path
+      end
     else
       @message = I18n.t('inscription_success.with_widget')
       @again = false
@@ -118,7 +131,7 @@ class ChildrenController < ApplicationController
   end
 
   def parent1_params
-    params.require(:child).permit(parent1_attributes: %i[letterbox_name address postal_code city_name first_name last_name phone_number gender degree_level_at_registration degree_country_at_registration preferred_channel])[:parent1_attributes]
+    params.require(:child).permit(parent1_attributes: %i[letterbox_name address postal_code city_name first_name last_name phone_number gender degree_level_at_registration degree_country_at_registration address_supplement email])[:parent1_attributes]
   end
 
   def parent2_params
